+++ melos.yaml
@@ -38,6 +38,10 @@ scripts:
   format:
     run: melos exec -- dart format --line-length=80 MELOS_PACKAGE_PATH
 
+  # Analyze code
+  analyze:
+    run: melos exec -- dart analyze MELOS_PACKAGE_PATH
+
   # Run build and format code
   brunner_format:
     run: melos run brunner && melos run format
